refactor(tui): trim workspace facade and settle cross-owner types (W5 CR)
