+++ src/tui/workspace/index.ts
@@ -1,34 +1,12 @@
 /**
- * Workspace domain facade. This is the single public import point for the
- * Workspace page outside the domain: App, the page view, and tests import the
- * public surface from here; internal owners do not reach through each other's
- * state. The surface is the complete set of exports external consumers may
- * use; W5 removes the root compat module after callers switch to this path.
+ * Workspace domain facade — the single public import point for the Workspace
+ * page outside the domain. App and the cross-domain tests import the
+ * controller factory from here; internal owners, the page view, and
+ * domain-focused tests import their target modules directly. Only the
+ * consumed surface is re-exported: adding an export here is a public-surface
+ * decision, not an automatic barrel.
  */
 export { createWorkspaceController } from "./controller";
 export { resetStaleHorizontalScroll } from "./buffer-owner";
 export type { WorkspaceController } from "./controller";
-export type {
-  ShellPage,
-  WorkspaceFocusRegion,
-  ViewerScrollEdge,
-  EditorStatusTone,
-} from "./types";
-export {
-  classifyFile,
-  OVERSIZED_BYTES,
-  PREVIEW_LINE_CAP,
-  scanDirectory,
-  flattenVisibleTree,
-  buildFileIndex,
-  matchFiles,
-  readFilePreview,
-  statEntry,
-} from "./tree-data";
-export type {
-  WorkspaceFileKind,
-  WorkspaceTreeNode,
-  WorkspaceVisibleRow,
-  WorkspaceFilePreview,
-} from "./tree-data";
-export { assertProjectRelativePath } from "./paths";
+export type { EditorStatusTone } from "./types";
